Changed player flipping and Attacking.
Player flipping - Instead of changing the players transform.scale.x to -/+ 1. There is now two attack hitboxes one for left and one for right. Then based on the faced direction do hit detection for left or right. The player sprite is flipped based on the direction it should face. I changed it because flipping the transform gave warnings that it may cause collision to work not as intended. This way gives no warnings.
Attacking - Previously I had it just enable and disable the hitbox with a coroutine. Now I use Physics.OverlapBox to see which colliders overlap. This way seems better as Coroutines can produce unintended results sometimes.